refactor: rename kubernetes-context-state to contexts-manager Signed-off-by: Philippe Martin <phmartin@redhat.com>
